Implemented protobuf into map factory
